Move APIs to Language and NamespacedLanguage
